requirements.txt
flask==2.0.1
Flask-Classful==0.14.1
wandb==0.12.4
memory-profiler==0.58.0
matplotlib==3.3.4
tensorflow==2.7
tensorflow-probability==0.15.0
tensorflow-addons==0.16.1
torch==1.9.0
torchvision==0.10.0
scipy==1.5.4
sklearn
tqdm
numpy~=1.18.5
pytest~=7.1.2
PyYAML~=5.4.1
jasscpp~=0.0.1
gym~=0.18.3
ray~=1.7.0
setuptools~=58.0.4